+++ ci/integration_tests/test_channel.py
@@ -1,9 +1,13 @@
 import os
+import socket
 from subprocess import check_output
 from unittest import skipUnless
 
+
+from ssh2.session import Session
 from ssh2.channel import Channel
 from ssh2.exceptions import SocketSendError
+from ssh2.session import LIBSSH2_METHOD_COMP_CS, LIBSSH2_FLAG_COMPRESS
 
 from .base_test import SSH2TestCase
 
@@ -193,3 +197,20 @@ class ChannelTestCase(SSH2TestCase):
         self.assertTrue(chan.execute(self.cmd) == 0)
         size, data = chan.read()
         self.assertTrue(size > 0)
+
+    def test_execute_with_session_compression(self):
+        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        sock.connect((self.host, self.port))
+        session = Session()
+        session.flag(LIBSSH2_FLAG_COMPRESS)
+        algs = session.supported_algs(LIBSSH2_METHOD_COMP_CS)
+        self.assertTrue('zlib' in algs)
+        session.handshake(sock)
+        session.userauth_publickey_fromfile(
+            self.user, self.user_key)
+        chan = session.open_session()
+        self.assertTrue(chan is not None)
+        self.assertTrue(chan.execute(self.cmd) == 0)
+        size, data = chan.read()
+        lines = [line.decode('utf-8') for line in data.splitlines()]
+        self.assertTrue(lines, [self.resp])