+++ ssh2/session.pyx
@@ -1,5 +1,5 @@
 # This file is part of ssh2-python.
-# Copyright (C) 2017-2020 Panos Kittenis
+# Copyright (C) 2017-2025 Panos Kittenis
 #
 # This library is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -53,6 +53,7 @@ LIBSSH2_HOSTKEY_TYPE_RSA = c_ssh2.LIBSSH2_HOSTKEY_TYPE_RSA
 LIBSSH2_HOSTKEY_TYPE_DSS = c_ssh2.LIBSSH2_HOSTKEY_TYPE_DSS
 
 
+## Method types and definitions
 cdef class MethodType:
     def __cinit__(self, value):
         self.value = value
@@ -68,6 +69,19 @@ LIBSSH2_METHOD_COMP_CS = MethodType(c_ssh2.LIBSSH2_METHOD_COMP_CS)
 LIBSSH2_METHOD_COMP_SC = MethodType(c_ssh2.LIBSSH2_METHOD_COMP_SC)
 LIBSSH2_METHOD_LANG_CS = MethodType(c_ssh2.LIBSSH2_METHOD_LANG_CS)
 LIBSSH2_METHOD_LANG_SC = MethodType(c_ssh2.LIBSSH2_METHOD_LANG_SC)
+
+
+## Flag types and definitions
+cdef class FlagType:
+    def __cinit__(self, value):
+        self.value = value
+
+
+LIBSSH2_FLAG_SIGPIPE = FlagType(c_ssh2.LIBSSH2_FLAG_SIGPIPE)
+LIBSSH2_FLAG_COMPRESS = FlagType(c_ssh2.LIBSSH2_FLAG_COMPRESS)
+LIBSSH2_FLAG_QUOTE_PATHS = FlagType(c_ssh2.LIBSSH2_FLAG_QUOTE_PATHS)
+LIBSSH2_FLAG_SK_PRESENCE_REQUIRED = FlagType(c_ssh2.LIBSSH2_SK_PRESENCE_REQUIRED)
+LIBSSH2_FLAG_SK_VERIFICATION_REQUIRED = FlagType(c_ssh2.LIBSSH2_SK_VERIFICATION_REQUIRED)
 
 
 cdef void kbd_callback(const char *name, int name_len,
@@ -447,6 +461,41 @@ cdef class Session:
             rc = c_ssh2.libssh2_session_block_directions(
                 self._session)
         return rc
+
+    def flag(self, FlagType flag, enabled=True):
+        """
+        Enable/Disable flag for session.
+
+        Flag must be one of :py:class:`ssh2.session.LIBSSH2_FLAG_SIGPIPE`
+          or :py:class:`ssh2.session.LIBSSH2_FLAG_COMPRESS`.
+
+        Flags *must* be set before :py:func:`Session.handshake` is called for the library to use them.
+
+        :py:class:`ssh2.session.LIBSSH2_FLAG_SIGPIPE` - Library will not block SIGPIPE signal from triggering from the
+          socket used. Meaning if the socket connection is terminated unexpectedly, using library functions will
+          trigger a SIGPIPE signal from the associated socket. Default is off.
+
+        :py:class:`ssh2.session.LIBSSH2_FLAG_COMPRESS` - Library will enable compression for the session.
+          Default is off.
+
+        Use `Session.supported_algs(LIBSSH2_METHOD_COMP_CS)` to get a list of supported compression algorithms after
+        enabling compression, if any.
+
+        Default is to enable the flag - `enabled=True`.
+
+        Set `enabled=False` to disable a previously enabled flag.
+
+        :raises ValueError: On incorrect :py:class:`ssh2.session.FlagType` passed.
+        :returns: None
+        """
+        cdef int rc
+        cdef bint value = enabled
+        if not flag in (LIBSSH2_FLAG_SIGPIPE, LIBSSH2_FLAG_COMPRESS):
+            raise ValueError("Provided flag must be one of LIBSSH2_FLAG_SIGPIPE or LIBSSH2_FLAG_COMPRESS - got %s",
+                             flag)
+        with nogil:
+            rc = c_ssh2.libssh2_session_flag(self._session, flag.value, value)
+        handle_error_codes(rc)
 
     def forward_listen(self, int port):
         """Create forward listener on port.