+++ tests/test_session.py
@@ -1,4 +1,7 @@
-from ssh2.session import Session, LIBSSH2_METHOD_HOSTKEY
+from ssh2.session import (Session, LIBSSH2_METHOD_HOSTKEY, LIBSSH2_FLAG_SIGPIPE, LIBSSH2_FLAG_COMPRESS,
+                          LIBSSH2_FLAG_QUOTE_PATHS, LIBSSH2_FLAG_SK_PRESENCE_REQUIRED,
+                          LIBSSH2_FLAG_SK_VERIFICATION_REQUIRED,
+                          )
 
 from .base_test import SSH2TestCase
 
@@ -12,3 +15,12 @@ class SessionTestCase(SSH2TestCase):
         session = Session()
         methods = session.methods(LIBSSH2_METHOD_HOSTKEY)
         self.assertIsNone(methods)
+
+    def test_flags(self):
+        session = Session()
+        for flag in [LIBSSH2_FLAG_SIGPIPE, LIBSSH2_FLAG_COMPRESS]:
+            session.flag(flag)
+            session.flag(flag, enabled=False)
+        for bad_flag in (LIBSSH2_FLAG_QUOTE_PATHS, LIBSSH2_FLAG_SK_PRESENCE_REQUIRED,
+                         LIBSSH2_FLAG_SK_VERIFICATION_REQUIRED):
+            self.assertRaises(ValueError, session.flag, bad_flag)